Bugfix. Ttimestamp and tx position are numbers
On the previous refactoring we have introduced a bug where the block
timestamp is left in hex format, also transaction position was being
decoded as string and not as number. This resulted in producing Kafka
data that consumers could not read.
The problems are now fixed along with tests added. For the block
timestamp problem a new test is added, while for the transaction
position the previous test is slightly modified so that it hit the
problem.